refactor(services): migrate coordination_resolver to orchestra/coordination
- Move coordination_resolver.py → orchestra/coordination.py
- Update imports in domain/qualify_gate.py and services/check/service.py
- Update lazy import mapping in services/__init__.py

Part of issue #2586: Phase 3-cleanup
